implement returning proper objects (mocked user)

+++ src/main/java/pl/edu/agh/tai/dilemmasask/api/controller/CommentController.java
@@ -3,6 +3,8 @@ package pl.edu.agh.tai.dilemmasask.api.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.edu.agh.tai.dilemmasask.api.model.Comment;
+import pl.edu.agh.tai.dilemmasask.api.model.User;
 import pl.edu.agh.tai.dilemmasask.api.repository.*;
 
 @RestController
@@ -12,10 +14,24 @@ public class CommentController {
     public CommentController(CommentRepository commentRepository) {
         this.commentRepository = commentRepository;
     }
+    private final static User mockUser = new User("mock");
 
     @DeleteMapping("/comments/{commentId}")
     private ResponseEntity deleteComment(@PathVariable Long commentId){
-        commentRepository.deleteById(commentId);
-        return ResponseEntity.status(HttpStatus.OK).build();
+        if (isUserAllowedToDeleteComment(mockUser, commentId)) {
+            commentRepository.deleteById(commentId);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+    }
+
+    private boolean isUserAllowedToDeleteComment(User user, Long commentId) {
+        Comment comment = commentRepository.findById(commentId).orElse(null);
+        if(comment==null){
+            return false;
+        }
+        return comment.getAuthor().equals(user);
     }
 }
